did work

+++ src/utils/cloudinary.ts
@@ -1,5 +1,10 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs"
+import dotenv from "dotenv"
+
+dotenv.config({
+    path: "./.env"
+})
 
 cloudinary.config({
     cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
@@ -15,6 +20,18 @@ export const uploadToCloudinary = async (filePath: string) => {
         resource_type: "auto",
     })
 
+    if (!file) {
+        throw new Error('Failed to upload file to Cloudinary');
+    }
+
     fs.unlinkSync(filePath);
     return file;
+}
+
+export const deleteFromCloudinary = async (fileID: string) => {
+    if (!fileID) {
+        throw new Error('File ID is required');
+    }
+
+    await cloudinary.uploader.destroy(fileID)
 }
